pyproject.toml
[project]
name = "articleflow"
version = "0.1.0"
description = "automation article task manager"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "dotenv>=0.9.9",
    "nltk>=3.9.2",
    "openai>=2.0.0",
    "pandas>=2.3.3",
    "python-docx>=1.2.0",
    "requests>=2.32.5",
    "ruff>=0.13.2",
    "scikit-learn>=1.7.2",
    "streamlit>=1.50.0",
    "xmltodict>=1.0.2",
]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []
fixable = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["my_package"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
